requirements.txt
algoliasearch-django>=3.0,<4.0'
django
djangorestframework
djangorestframework-simplejwt
pyyaml
django-cors-headers
requests